only create tables/load data at last crdb replica

+++ src/store/cockroachdb/server.cc
@@ -171,30 +171,31 @@ void Server::Load(const string &key, const string &value,
 
 void Server::CreateTable(const std::string &table_name, const std::vector<std::pair<std::string, std::string>> &column_data_types, 
       const std::vector<uint32_t> &primary_key_col_idx) {
-
-   
-  std::string sql_statement("CREATE TABLE IF NOT EXISTS");
-  sql_statement += " " + table_name;
-  UW_ASSERT(!column_data_types.empty());
-  sql_statement += " (";
-  for (auto &[col, type] : column_data_types) {
-    sql_statement += col + " " + type + ", ";
-  }
-  if (!primary_key_col_idx.empty()) {
-    sql_statement += "PRIMARY KEY ";
-    sql_statement += "(";
-    for (auto &p_idx : primary_key_col_idx) {
-      sql_statement += (column_data_types[p_idx].first + ", ");
+  // Only the last node creates the table
+  if (id == numGroups * config.n - 1) {
+    std::string sql_statement("CREATE TABLE IF NOT EXISTS");
+    sql_statement += " " + table_name;
+    UW_ASSERT(!column_data_types.empty());
+    sql_statement += " (";
+    for (auto &[col, type] : column_data_types) {
+      sql_statement += col + " " + type + ", ";
     }
-    sql_statement.resize(sql_statement.size() - 2);  // remove trailing ", "
-    sql_statement += ")";
-  } else {
-    sql_statement.resize(sql_statement.size() - 2);  // remove trailing ", "
-  }
-  sql_statement += ");";
+    if (!primary_key_col_idx.empty()) {
+      sql_statement += "PRIMARY KEY ";
+      sql_statement += "(";
+      for (auto &p_idx : primary_key_col_idx) {
+        sql_statement += (column_data_types[p_idx].first + ", ");
+      }
+      sql_statement.resize(sql_statement.size() - 2);  // remove trailing ", "
+      sql_statement += ")";
+    } else {
+      sql_statement.resize(sql_statement.size() - 2);  // remove trailing ", "
+    }
+    sql_statement += ");";
 
-  Notice("Create Table: %s", table_name.c_str());
-  Server::exec_sql(sql_statement);
+    Notice("Create Table: %s", table_name.c_str());
+    Server::exec_sql(sql_statement);
+  }
 }
 
 void Server::LoadTableData(const std::string &table_name, const std::string &table_data_path, 
@@ -202,32 +203,34 @@ void Server::LoadTableData(const std::string &table_name, const std::string &tab
   // Syntax based on: https://www.cockroachlabs.com/docs/stable/import-into.html
   // https://www.cockroachlabs.com/docs/v22.2/use-a-local-file-server
   // data path should be a url
-  std::string file_server_path = std::filesystem::path(table_data_path).remove_filename();
-  std::string file_server_cmd = "cd " + file_server_path + "; python3 -m http.server 3000 &";
-  int status = system(file_server_cmd.c_str());
-  Notice("File server started on port 3000");
+  if (id == numGroups * config.n - 1) {
+    std::string file_server_path = std::filesystem::path(table_data_path).remove_filename();
+    std::string file_server_cmd = "cd " + file_server_path + "; python3 -m http.server 3000 &";
+    int status = system(file_server_cmd.c_str());
+    Notice("File server started on port 3000");
 
-  std::string table_file_name = std::filesystem::path(table_data_path).filename();
-  std::string table_column_name;
-  for (auto &[col, type] : column_names_and_types) {
-    table_column_name += "" + col + ",";
+    std::string table_file_name = std::filesystem::path(table_data_path).filename();
+    std::string table_column_name;
+    for (auto &[col, type] : column_names_and_types) {
+      table_column_name += "" + col + ",";
+    }
+    table_column_name.pop_back();
+
+    Notice("Loading table from path: %s", table_data_path.c_str());
+
+    std::string copy_table_statement_crdb = fmt::format(
+        "IMPORT INTO {0} ({1}) CSV DATA "
+        "(\'http://localhost:3000/{2}\') "
+        "WITH skip = \'1\'",
+        table_name, table_column_name,
+        table_file_name);  // FIXME: does one need to specify column names?
+                          // Target columns don't appear to be enforced
+
+    Notice("Load Table Data for table: %s", table_name.c_str());
+    exec_sql(copy_table_statement_crdb);
+    stats.Increment("TablesLoaded", 1);
+    system("pkill -f \"python3 -m http.server\"");
   }
-  table_column_name.pop_back();
-
-  Notice("Loading table from path: %s", table_data_path.c_str());
-
-  std::string copy_table_statement_crdb = fmt::format(
-      "IMPORT INTO {0} ({1}) CSV DATA "
-      "(\'http://localhost:3000/{2}\') "
-      "WITH skip = \'1\'",
-      table_name, table_column_name,
-      table_file_name);  // FIXME: does one need to specify column names?
-                         // Target columns don't appear to be enforced
-
-  Notice("Load Table Data for table: %s", table_name.c_str());
-  exec_sql(copy_table_statement_crdb);
-  stats.Increment("TablesLoaded", 1);
-  system("pkill -f \"python3 -m http.server\"");
 }
 
 void Server::CreateIndex(
